+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Process class that serves as a base for processes the kernel can run.
 - Added LinkNetwork class that handles logic concerning multiple StructureLink objects in a room.
 - Added Relations class that manages relations with other players.
+- Added RoomIntel class that replaces direct accesses to data in `Memory.rooms[roomName].intel`.
 - Added creep role "gift" that takes excess resources and runs them around the map for other players to hunt.
 - Spawn reserver creeps for rooms that are deemed "safe" by the room planner, because they cannot be accessed from outside our empire.
 
@@ -23,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Several function have been refactored for better readability and to reduce duplication.
 - We no longer create a new instace of the Logger class for almost every log message. Instead, the Hivemind class has a factory method for getting a Logger.
 - Expansions to other rooms will now avoid being close to other players, and prefer rooms that have many energy sources in adjacent rooms but few exit sides.
+- Expansions are now taken up to 7 rooms away (up from 5).
 - Military creeps will now attack unowned structures if a flag has been placed directly on it.
 - Remote mining will try not to run paths through rooms owned or reserved by other players.